Change shebang line to use /usr/bin/env
If you run the program in a virtual environment, then you want
to use the python interpreter inside that virtual environment.
Change the shebang line from #!/usr/bin/python3 (the system-wide
interpreter) to #!/usr/bin/env python3 (the first python3
interpreter found in the PATH).

This fixes a problem when running on bird-slideshow.py on old
distributions of Linux (specifically on Ubuntu 18.04), where
a virtual environment has to be created, because the PIL module
available for the distribution is too old (version 5.1.0) to
have the ImageTk object.

This change fixes an import error, using the system interpreter
and (thus) the old PIL module:
Traceback (most recent call last):
  File "./bird-slideshow.py", line 16, in <module>
    from PIL import Image, ImageTk
ImportError: cannot import name 'ImageTk'

Signed-off-by: Tim Bird <tim.bird@sony.com>