Add support for ssh paths
A source line in the config can now indicate an 'ssh path', which
has the format:
  ssh:{user}[:{password}]@{hostname}:{path}

An example is: ssh:ssuser:password@example-host.com:/home/ssuser/Pictures

This uses the name (and optional password) to get a directory
listing of the indicated directory on the host.  Image files from this
listing are downloaded using 'scp' into the cache directory, and
displayed as part of the slideshow.

Signed-off-by: Tim Bird <tim.bird@sony.com>